add spec when p is given both argument and block
I decide to change the behavior of p in that case:

p('title'){ titled_block } returns only the result of block.

Although not documented on the spec, p('argument'){ block } returns the Array of the arguments and the result of the block.

I think it's more useful if we could use the argument of p just as the description of the block